requirements.txt
oci
hydra-core
hydra_colorlog
setuptools